crowbar: Create helper module to validate states
